Version the git_sync structure as v1alpha1
