+++ crates/stackable-operator/src/crd/mod.rs
@@ -5,6 +5,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 pub mod authentication;
+pub mod git_sync;
 pub mod listener;
 pub mod s3;
